Switched from YAML database to SQL with Sequel
Why: Before this commit our database was a simple YAML file instead of a
full fledged DB. This commit was a big refactor that reworked all of the
methods to work with a SQL backend and the sequel wrapper for ease of
use.

*

This change addresses the need by: Changing all methods that were
touching the old YAML db and replacing it with Sequel methods.

*